Transactions database hook (#72)
* Added rough db connection

* Updated the transactions page

* some light work

+++ Website/demo/Models/Home/Transactions.cs
@@ -12,7 +12,7 @@ namespace demo.Models
 
         public string processing_date { get; set; }
 
-        public int balance { get; set; }
+        public double balance { get; set; }
 
         public string deposit_withdrawal { get; set; }
